requirements.txt
# Django & Frameworks
Django==5.2.6
djangorestframework==3.16.1
django-ninja==1.4.3
django-cors-headers==4.8.0
django-filter==25.1
django-environ==0.12.0

# Auth & JWT
djangorestframework-simplejwt==5.5.1

# Documentation
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1

# Server
gunicorn==23.0.0

# Database
mysqlclient==2.2.7

# Data Processing
pandas==2.3.3
openpyxl==3.1.5
xlrd==2.0.1
PyYAML==6.0.2
tzdata==2025.2

# Validation
pydantic==2.11.9
pydantic-core==2.33.2
typing_extensions==4.15.0

# Testing (Standardization)
pytest==8.3.4
pytest-django==4.9.0
